Dices and Slots are now provably fair.
Dices and slots are in the end provably fair!

How it's it working?

The server is creating two secrets for each player, they are shown as hashes before the player spins. And there is a secret that it's provided by the player.

A sha512-hmac function is executed with the server secret as data and player secret as key.

Then the first 5 characters from the hex are transformed into integers. If it's lower than 999999 for dices or 1000000 for slots it is divided by 100 for dices and 10000 for slots.

If not the next 5 characters and so on. If all of the 125 characters are over the numbers for each, which is almost impossible, the dice roll will be 4500 and the slots will be an orange.